A lot again SharedPreferences: Try to use this method for storing extra data about the user on the device longer than just one session. Don't quite know yet how to format the data, but I will figure something out.
Remote Config
Enabled remote Config for the current semester. sometimes it still gets the "default" value, but I think, that was just a debug error.

Login
Completely redesigned the login and register methods. There are more sites in the dialog now and the code is hopefully easier to read.

StartActivity
Changed the way the signed in user should get rights. The signed in user has "admin" rights, if its uid is to be found in the current chargen document in the firestore database. also a signed in user has admin rights, if its uid is in the "Editor.super-admin" document.